+++ common/src/main/webapp/sass/explore.scss
@@ -639,7 +639,7 @@ thead tr th.sticky-header {
   justify-content: center;
 
   .p-component.p-dropdown {
-    width: 50em;
+    width: 30em;
   }
 }
