+++ model/shared/src/main/scala/explore/model/package.scala
@@ -58,7 +58,7 @@ type ObservationList            = SortedMap[Observation.Id, Observation]
 type GroupList                  = SortedMap[Group.Id, Group]
 type ConstraintGroupList        = SortedMap[ObsIdSet, ConstraintSet]
 type SchedulingGroupList        = SortedMap[ObsIdSet, List[TimingWindow]]
-type ObservingModeGroupList     = SortedMap[ObsIdSet, Option[EffectiveObservingMode]]
+type ObservingModeGroupList     = SortedMap[ObsIdSet, Option[ObservingModeSummary]]
 type ObsAttachmentList          = SortedMap[ObsAtt.Id, ObsAttachment]
 type ObsAttachmentAssignmentMap = Map[ObsAtt.Id, SortedSet[Observation.Id]]
 type ProgramInfoList            = SortedMap[Program.Id, ProgramInfo]
