+++ project/Versions.scala
@@ -24,12 +24,12 @@ object Versions {
   val lucumaCore             = "0.107.1"
   val lucumaCatalog          = "0.48.6"
   val lucumaITC              = "0.22.6"
-  val lucumaReact            = "0.72.0"
+  val lucumaReact            = "0.73.0"
   val lucumaRefined          = "0.1.3"
   val lucumaSchemas          = "0.107.1"
   val lucumaOdbSchema        = "0.16.1"
   val lucumaSSO              = "0.6.27"
-  val lucumaUI               = "0.123.0"
+  val lucumaUI               = "0.123.0-4-7b132e8-20241118T205522Z-SNAPSHOT"
   val monocle                = "3.3.0"
   val mouse                  = "1.3.2"
   val mUnit                  = "1.0.2"
